requirements.txt
appdirs~=1.4.4
configparser~=5.2.0
gower~=0.0.5
iteround~=1.0.3
liac-arff>=2.5.0
matplotlib~=3.4.2
mininet~=2.3.0.dev6
mysqlclient~=2.0.3
networkx~=2.8
numpy>=1.22.0
pandas~=1.2.4
paramiko~=2.7.2
pexpect~=4.8.0
PyPika~=0.48.8
python-javabridge>=4.0.0
python-weka-wrapper3>=0.2.5
seaborn~=0.11.1
setuptools>=57.0.0
scipy>=1.6.3
sympy~=1.8
tabulate~=0.8.9
z3-solver>=4.8.12.0